+++ projects/core/src/main/kotlin/site/siredvin/peripheralium/ext/Set.kt
@@ -1,0 +1,3 @@
+package site.siredvin.peripheralium.ext
+
+infix fun <T> Set<T>.xor(that: Set<T>): Set<T> = (this - that) + (that - this)